+++ changelog.md
@@ -4,6 +4,20 @@
 - Added Bedrock Skin (wip)
 - Added Ram
 - Added Ram sound effect
+- Particle effects (For now angry villager, until I make my own) to Taunted enemies
+
+### Changed
+- Centered taunt spell png
+- Changed custom lang to ISS
+- Revamped sworn protector
+
+## [0.4] - 2026-07-
+
+### Added
+- Added Bedrock Skin (wip)
+- Added Ram
+- Added Ram sound effect
+- Particle effects (For now angry villager, until I make my own) to Taunted enemies
   
 ### Changed
 - Centered taunt spell png